plugins/dwd/requirements.txt
Flask==2.1.3
flask_restful
flask-mongoengine
flask_mongorest
flask-session
mimerender
cleancat
mongoengine==0.23.1
Flask-WTF
wheel
bcrypt
uuid
redis
requests
awesome-slugify
pymongo
Werkzeug==2.2.2
deep-translator
translate
python-dotenv
datetime
mongodb-migrations


